fix: reduce disk sizes
We reduced recently the resources and workload of the benchmarks, but not the disk sizes. This commit reduces
the disk size by 1/4. The disk size for zeebe is now similar to the cluster plan G3-S, and it is still serving
a similar workload.

This should help to reduce further our costs of running benchmarks