fix: use correct ES uri
Due to ES chart migration, the name of the pods, pvcs etc changed.
In consequence we need to use a different URI to connect to the elastic nodes.
This commit adjustes the configured URI for the prometheus es exporter, and
uses the templating logic, to resolve the release name.

This is only possible if the value is consumed via tpl, which is the case

https://github.com/prometheus-community/helm-charts/blob/801b161283505e1d6d63129fea759126e33e0d18/charts/prometheus-elasticsearch-exporter/templates/deployment.yaml#L88