refactor: load config as additional spring config
Use SPRING_CONFIG_ADDITIONALLOCATION to load an additional configuration on top
of the provided configuration by the parent platform helm chart.
https://docs.spring.io/spring-boot/reference/features/external-config.html#features.external-config.files.optional-prefix

Renaming the old configmap and application.yaml file to benchmark-config.yaml to not clash
with the mounting of the default application.yaml from the parent chart.

Allows to load two configurations, and override additional properties.