fix: reduce capacity of workers
Most of the workers have a completion delay of 300ms, and per default 10 threads.
This means per second they are able to complete ~30 in one second. As
the timeout is calculated based on the completion delay (times 6 = 1800 ms). We
have a high chance that jobs time out until we reach them.

Reducing the capacity should remove this risk or reduce the chance.